Modify Candidate and CandidateProfile
Most personal information for Candidates was moved to a separate model,
CandidateProfile. However, profession had to be kept in the users table
to simplify joining tables. Also, nested attributes don't go well with
virtual attribute profession_name that is used in this case.
This commit modifies an already-commited migration, but otherwise a
migration adding profession_id back to the users table would be
conflicting with an existing migration.